Add a development container for VS Code
This creates a development environment with a FoundationDB server and
an erlfdb client in two containers, sharing a network through Docker
Compose.

It uses the FDB image published to Docker Hub for the FDB container, and
downloads the FDB client packages from foundationdb.org to provide the
development headers and libraries. Once the Docker Compose setup is
running, VS Code executes the `create_cluster_file.bash` script to write
down a cluster file containing the IP address in the compose network
where the FDB service can be found.