Create a DB if the test cluster doesn't have one
The FoundationDB image does not always ship with a configured database,
so we'll attempt to create one when initializing the test suite.

Note that this can cause the FDB server to lose data if it is in a
misconfigured state. We only run this command on test clusters, but we
ought to warn users not to run the test suite against a production FDB
cluster all the same.